Module Catapult_sqlite

Backend that writes directly to a Sqlite database.

Creating the database is done via Writer.create or Writer.with_. Then Backend can be used to turn this Writer.t into a Tracing collector.

module Writer : sig ... end

DB writer.

module Backend : sig ... end

val backend_of_writer : Writer.t -> Catapult.backend

val trace_collector_of_writer : Writer.t -> Trace_core.collector

Turn a writer into a Trace collector.

val with_ : ?sync:[ `FULL | `NORMAL | `OFF ] -> ?append:bool -> ?file:string -> ?trace_id:string -> ?dir:string -> unit -> (unit -> 'a) -> 'a

with_ () f runs f() in a scope where a connection to a Sqlite DB has been established and is used to store tracing events emitted from within f ().